+++ examples/docs/app/[[...mdxPath]]/page.jsx
@@ -1,7 +1,7 @@
 /* eslint-env node */
 /* eslint-disable react-hooks/rules-of-hooks -- false positive, useMDXComponents/useTOC are not react hooks */
 
-import { getPagesPaths, importPage } from 'nextra'
+import { getPagesPaths, importPage } from 'nextra/pages'
 import { useMDXComponents } from 'nextra-theme-docs'
 
 export async function generateStaticParams() {
