+++ packages/nextra/package.json
@@ -36,6 +36,10 @@
       "import": "./dist/server/page-map.js",
       "types": "./dist/server/page-map.d.ts"
     },
+    "./pages": {
+      "import": "./dist/server/pages.js",
+      "types": "./dist/server/pages.d.ts"
+    },
     "./*": {
       "import": "./dist/client/*.js",
       "types": "./dist/client/*.d.ts"
@@ -46,6 +50,9 @@
     "*": {
       "page-map": [
         "./dist/server/page-map.d.ts"
+      ],
+      "pages": [
+        "./dist/server/pages.d.ts"
       ],
       "compile": [
         "./dist/server/compile.d.ts"
